chore: complete refactor
